Add `-Wl,-export-dynamic` to the amalgamated build to allow it to dynamically load `lighttpd` `mod_*.so`s, even C ones.
The previous `rust/` transpile turned the dynamically loaded modules into statically linked Rust modules (I think),
so this wasn't an issue before, but since the amalgamated build is for one binary (`lighttpd`), this wasn't done.
Perhaps it could be done, too, but for now this works and allows C and Rust to call each other over dynamically loaded FFI.

Link commands aren't saved in `compile_commands.json`, so the link flags are a bit trickier to hunt down and are currently hardcoded.

Since a few modules are needed for running `lighttpd`, `amalgamate.sh` now builds them, but after generating the `compile_commands.json`.